Add javascript build step using rollup
