+++ package.json
@@ -1,7 +1,12 @@
 {
-    "dependencies": {
+    "type": "module",
+    "devDependencies": {
+        "@rollup/plugin-swc": "^0.4.1",
+        "@swc/core": "^1.15.47",
         "@tailwindcss/cli": "^4.2.2",
-        "tailwindcss": "^4.2.2"
+        "rollup": "^4.62.4",
+        "tailwindcss": "^4.2.2",
+        "@rollup/plugin-terser": "^1.0.0"
     },
     "resolutions": {
         "picomatch": "^2.3.2"
@@ -9,6 +14,9 @@
     "scripts": {
         "dev": "yarn watch",
         "watch": "tailwindcss --input ./resources/css/mailbook.css --output ./resources/dist/mailbook.css --watch",
-        "build": "tailwindcss --input ./resources/css/mailbook.css --output ./resources/dist/mailbook.css --minify"
+        "watch:js": "yarn build:js --watch",
+        "build": "yarn build:css && yarn build:js",
+        "build:css": "tailwindcss --input ./resources/css/mailbook.css --output ./resources/dist/mailbook.css --minify",
+        "build:js": "rollup --config rollup.config.js"
     }
 }
