+++ resources/js/index.js
@@ -1,0 +1,7 @@
+import {collapsableGroups} from "./collapsableGroups";
+import {sidebarScrolling} from "./sidebarScrolling";
+import {languageSelector} from './languageSelector';
+
+collapsableGroups();
+sidebarScrolling();
+languageSelector();
